Make pipeline tests more configurable and fix CircleCI config
