+++ .circleci/config.yml
@@ -73,6 +73,20 @@ jobs:
                   keys:
                       - repo-{{ .Environment.CIRCLE_SHA1 }}
             - run: yarn test:generate_docs:circleci
+    test-pipeline:
+        docker:
+            - image: circleci/node:9
+            - image: postgres:11-alpine
+        working_directory: ~/repo
+        steps:
+            - restore_cache:
+                  keys:
+                      - repo-{{ .Environment.CIRCLE_SHA1 }}
+            - run: ZEROEX_DATA_PIPELINE_TEST_DB_URL='postgresql://postgres@localhost/postgres' yarn wsrun test:circleci @0x/pipeline
+            - save_cache:
+                  key: coverage-pipeline-{{ .Environment.CIRCLE_SHA1 }}
+                  paths:
+                      - ~/repo/packages/pipeline/coverage/lcov.info
     test-rest:
         docker:
             - image: circleci/node:9
@@ -98,7 +112,6 @@ jobs:
             - run: yarn wsrun test:circleci @0x/web3-wrapper
             - run: yarn wsrun test:circleci @0x/utils
             - run: yarn wsrun test:circleci @0x/instant
-            - run: yarn wsrun test:circleci @0x/pipeline
             - save_cache:
                   key: coverage-abi-gen-{{ .Environment.CIRCLE_SHA1 }}
                   paths:
@@ -330,6 +343,9 @@ workflows:
                   requires:
                       - build
             - test-contracts-geth:
+                  requires:
+                      - build
+            - test-pipeline:
                   requires:
                       - build
             - test-rest:
