+++ packages/pipeline/package.json
@@ -9,7 +9,7 @@
         "test": "yarn run_mocha",
         "rebuild_and_test": "run-s build test",
         "test:circleci": "yarn test:coverage",
-        "run_mocha": "mocha --require source-map-support/register --require make-promises-safe lib/test/**/*_test.js lib/test/**/**/*_test.js --bail --exit --timeout 60000",
+        "run_mocha": "mocha --require source-map-support/register --require make-promises-safe lib/test/global_hooks.js lib/test/**/*_test.js lib/test/**/**/*_test.js --bail --exit --timeout 60000",
         "test:coverage": "nyc npm run test --all && yarn coverage:report:lcov",
         "coverage:report:lcov": "nyc report --reporter=text-lcov > coverage/lcov.info",
         "clean": "shx rm -rf lib",
