+++ packages/pipeline/test/entities/block_test.ts
@@ -2,21 +2,15 @@ import 'mocha';
 import 'reflect-metadata';
 
 import { Block } from '../../src/entities';
+import { createDbConnectionOnceAsync } from '../db_setup';
 import { chaiSetup } from '../utils/chai_setup';
 
-import { createDbConnectionOnceAsync, setUpDbAsync, tearDownDbAsync } from './db_setup';
 import { testSaveAndFindEntityAsync } from './util';
 
 chaiSetup.configure();
 
 // tslint:disable:custom-no-magic-numbers
 describe('Block entity', () => {
-    before(async () => {
-        await setUpDbAsync();
-    });
-    after(async () => {
-        await tearDownDbAsync();
-    });
     it('save/find a block', async () => {
         const connection = await createDbConnectionOnceAsync();
         const block = new Block();
